Support interpolations in regular double quoted strings. Remove curly bracket support.
This is a breaking change. The compiler source has quite lot of string literals (about 70) that include curly brackets that need to be properly quoted.

Rather than quoting them with \ escapes, I'd rather use double {{ }} quoting, so need to temporarily extend the old string literal tokenizer to support this. The compiler should then bootstrap, using literal strings for the first compile and interpolated strings for the subsequent.